+++ example/lib/screens/surface/progress_indicators.dart
@@ -6,7 +6,7 @@ import 'package:fluent_ui/fluent_ui.dart';
 import '../../widgets/card_highlight.dart';
 
 class ProgressIndicatorsPage extends StatefulWidget {
-  const ProgressIndicatorsPage({Key? key}) : super(key: key);
+  const ProgressIndicatorsPage({super.key});
 
   @override
   State<ProgressIndicatorsPage> createState() => _ProgressIndicatorsPageState();
@@ -35,9 +35,9 @@ class _ProgressIndicatorsPageState extends State<ProgressIndicatorsPage>
             'completion time is unknown.',
           ),
         ),
-        CardHighlight(
+        const CardHighlight(
           child: RepaintBoundary(
-            child: Row(children: const [
+            child: Row(children: [
               ProgressBar(),
               SizedBox(width: 20.0),
               ProgressRing(),