+++ lib/src/controls/form/combo_box.dart
@@ -16,7 +16,9 @@ const EdgeInsetsGeometry _kAlignedButtonPadding = EdgeInsetsDirectional.only(
   end: 15.0,
 );
 const EdgeInsets _kAlignedMenuMargin = EdgeInsets.zero;
-const EdgeInsets _kListPadding = EdgeInsets.only(top: _kMenuItemBottomPadding);
+const EdgeInsetsDirectional _kListPadding = EdgeInsetsDirectional.only(
+  top: _kMenuItemBottomPadding,
+);
 const kComboBoxRadius = Radius.circular(4.0);
 
 /// A builder to customize combo box buttons.
@@ -89,13 +91,13 @@ class _ComboBoxScrollBehavior extends FluentScrollBehavior {
 // The widget that is the button wrapping the menu items.
 class _ComboBoxItemButton<T> extends StatefulWidget {
   const _ComboBoxItemButton({
-    Key? key,
+    super.key,
     this.padding,
     required this.route,
     required this.buttonRect,
     required this.constraints,
     required this.itemIndex,
-  }) : super(key: key);
+  });
 
   final _ComboBoxRoute<T> route;
   final EdgeInsets? padding;
@@ -213,13 +215,13 @@ class _ComboBoxItemButtonState<T> extends State<_ComboBoxItemButton<T>> {
 
 class _ComboBoxMenu<T> extends StatefulWidget {
   const _ComboBoxMenu({
-    Key? key,
+    super.key,
     this.padding,
     required this.route,
     required this.buttonRect,
     required this.constraints,
     this.popupColor,
-  }) : super(key: key);
+  });
 
   final _ComboBoxRoute<T> route;
   final EdgeInsets? padding;
@@ -306,7 +308,7 @@ class _ComboBoxMenuState<T> extends State<_ComboBoxMenu<T>> {
                 namesRoute: true,
                 explicitChildNodes: true,
                 label: FluentLocalizations.of(context).dialogLabel,
-                child: DefaultTextStyle(
+                child: DefaultTextStyle.merge(
                   style: route.style,
                   child: ScrollConfiguration(
                     behavior: const _ComboBoxScrollBehavior(),
@@ -620,7 +622,7 @@ class _ComboBoxRoute<T> extends PopupRoute<_ComboBoxRouteResult<T>> {
 
 class _ComboBoxRoutePage<T> extends StatelessWidget {
   const _ComboBoxRoutePage({
-    Key? key,
+    super.key,
     required this.route,
     required this.constraints,
     required this.padding,
@@ -630,7 +632,7 @@ class _ComboBoxRoutePage<T> extends StatelessWidget {
     required this.capturedThemes,
     this.style,
     required this.popupColor,
-  }) : super(key: key);
+  });
 
   final _ComboBoxRoute<T> route;
   final BoxConstraints constraints;
@@ -695,9 +697,9 @@ class _ComboBoxItemContainer extends StatelessWidget {
   ///
   /// The [child] argument is required.
   const _ComboBoxItemContainer({
-    Key? key,
+    super.key,
     required this.child,
-  }) : super(key: key);
+  });
 
   /// The widget below this widget in the tree.
   ///
@@ -716,10 +718,7 @@ class _ComboBoxItemContainer extends StatelessWidget {
 }
 
 class _ContainerWithoutPadding extends InheritedWidget {
-  const _ContainerWithoutPadding({
-    Key? key,
-    required Widget child,
-  }) : super(key: key, child: child);
+  const _ContainerWithoutPadding({required super.child});
 
   static _ContainerWithoutPadding? of(BuildContext context) {
     return context.dependOnInheritedWidgetOfExactType<_ContainerWithoutPadding>();
@@ -740,11 +739,11 @@ class ComboBoxItem<T> extends _ComboBoxItemContainer {
   ///
   /// The [child] argument is required.
   const ComboBoxItem({
-    Key? key,
+    super.key,
     this.onTap,
     this.value,
-    required Widget child,
-  }) : super(key: key, child: child);
+    required super.child,
+  });
 
   /// Called when the combo box menu item is tapped.
   final VoidCallback? onTap;
@@ -809,7 +808,7 @@ class ComboBox<T> extends StatefulWidget {
   /// The [popupColor] argument specifies the background color of the
   /// combo box when it is open. If it is null, the default [Acrylic] color is used.
   const ComboBox({
-    Key? key,
+    super.key,
     required this.items,
     this.selectedItemBuilder,
     this.value,
@@ -830,7 +829,7 @@ class ComboBox<T> extends StatefulWidget {
     this.popupColor,
     // When adding new arguments, consider adding similar arguments to
     // ComboBoxFormField.
-  }) : super(key: key);
+  });
 
   /// The list of items the user can select.
   ///
@@ -1193,7 +1192,7 @@ class ComboBoxState<T> extends State<ComboBox<T>> {
       }
 
       placeholderIndex = items.length;
-      items.add(DefaultTextStyle(
+      items.add(DefaultTextStyle.merge(
         style: textStyle!.copyWith(color: theme.disabledColor),
         child: IgnorePointer(
           ignoringSemantics: false,
@@ -1223,7 +1222,7 @@ class ComboBoxState<T> extends State<ComboBox<T>> {
     }
 
     Widget result = Builder(builder: (context) {
-      return DefaultTextStyle(
+      return DefaultTextStyle.merge(
         style: isEnabled ? textStyle! : textStyle!.copyWith(color: theme.disabledColor),
         child: Container(
           padding: padding.resolve(Directionality.of(context)),